requirements.txt
flask==3.0.0
python-dotenv==1.0.1
google-genai==0.3.0
pymongo==4.6.1
bcrypt==4.1.2
reportlab==4.0.8
pillow>=10.2.0
opencv-python-headless>=4.10.0.84
urllib3==1.26.18